requirements.txt
lark>=1.1.9
PyYAML>=6.0.1
pytest>=7.4.0
